Minor fixes
- Allows for file selection (instead of directory selection) when selecting the PhysX SDK path.
- Prevents error when the user tries to export a mesh without simulated vertices. Drive paint is still necessary to define a physical mesh, even if no vertex is simulated.

+++ io_mesh_apx/utils.py
@@ -389,7 +389,7 @@ def getWeightArray(verts, vg):
             weights.append(vg.weight(i))
         except:
             weights.append(0)
-    return np.array(weights)
+    return np.array(weights, dtype = 'float64')
 
 def simValueType(type):
     match type:
